+++ src/components/Form/Fields/EmailInput/EmailInput.tsx
@@ -1,15 +1,15 @@
 import * as React from "react";
-import { IFormInputBaseWithValidation } from '../../../../models/IFormInput';
+import { IEmailInput } from '../../../../models/IFormInput';
 import { useFormContext } from 'react-hook-form';
 
-export const EmailInput: React.FunctionComponent<IFormInputBaseWithValidation> = ({ name, label, required, className, inlineLabel, disabled, placeholder, defaultValue, requiredValidationMessage }) => {
+export const EmailInput: React.FunctionComponent<IEmailInput> = ({ name, label, required, className, inlineLabel, disabled, placeholder, defaultValue, requiredValidationMessage, labelCol = 4, inputCol = 8 }) => {
 
     const { errors, register } = useFormContext();
 
     return (
         <div className={className + " form-group " + (inlineLabel ? "row" : "")}>
-            <label className={inlineLabel ? "col-4 col-form-label" : ""}>{label}:{required ? "*" : ""}</label>
-            <div className={inlineLabel ? "col-8" : ""}>
+            <label className={inlineLabel ? `col-${labelCol} col-form-label` : ""}>{label}:{required ? "*" : ""}</label>
+            <div className={inlineLabel ? `col-${inputCol}` : ""}>
                 <input type="email"
                     name={name}
                     id={name}