+++ src/components/Form/Fields/TextInput/TextInput.stories.tsx
@@ -4,60 +4,69 @@ import { useForm, FormContext } from 'react-hook-form';
 
 import { TextInput } from './TextInput';
 import { ITextInput } from '../../../../models/IFormInput';
+import { CustomSubmitButton, Form } from '../../Form';
 
 // Extend ITextInput to include formMethods for Storybook.
 interface StorybookITextInput extends ITextInput {
-  formMethods: any; // Replace 'any' with the actual type if you know it
+    formMethods: any; // Replace 'any' with the actual type if you know it
 }
 
 export default {
-  title: 'Form/Fields/TextInput',
-  component: TextInput,
+    title: 'Form/Fields/TextInput',
+    component: TextInput,
 } as Meta;
 
 const Template: Story<StorybookITextInput> = (args) => {
-  const methods = useForm();
+    const methods = useForm();
 
-  return (
-    <FormContext {...methods}>
-      <TextInput {...args} />
-    </FormContext>
-  );
+    const onSubmit = (data: any) => {
+        console.log('Form Submitted:', data);
+    };
+
+    return (
+        <Form {...methods} onSubmit={onSubmit}>
+            <TextInput {...args} />
+            <CustomSubmitButton>Test</CustomSubmitButton>
+        </Form>
+    );
 };
 
 export const Default = Template.bind({});
 Default.args = {
-  label: 'Default Label',
-  inlineLabel: true,
+    name: "default",
+    label: 'Default Label',
+    inlineLabel: true,
 };
 
 export const Disabled = Template.bind({});
 Disabled.args = {
-  label: 'Disabled Label',
-  inlineLabel: true,
-  disabled: true,
+    name: 'disabled',
+    label: 'Disabled Label',
+    inlineLabel: true,
+    disabled: true,
 };
 
 export const Readonly = Template.bind({});
 Readonly.args = {
-  label: 'Readonly Label',
-  name: 'readonly',
-  inlineLabel: true,
-  defaultValue: "Readonly",
-  readonly: true,
+    label: 'Readonly Label',
+    name: 'readonly',
+    inlineLabel: true,
+    defaultValue: "Readonly",
+    readonly: true,
 };
 
 export const WithPlaceholder = Template.bind({});
 WithPlaceholder.args = {
-  label: 'With Placeholder',
-  inlineLabel: true,
-  placeholder: 'Type here...',
+    label: 'With Placeholder',
+    name: 'withPlaceholder',
+    inlineLabel: true,
+    placeholder: 'Type here...',
 };
 
 export const RequiredField = Template.bind({});
 RequiredField.args = {
-  label: 'Required Field',
-  inlineLabel: true,
-  required: true,
-  requiredValidationMessage: 'This field is required',
+    label: 'Required Field',
+    name: 'required',
+    inlineLabel: true,
+    required: true
 };