+++ src/index.tsx
@@ -1,4 +1,4 @@
-﻿import { Form, Dropdown, ConditionalCheckboxList, SelectGroup, UnorderedList, TooltipItem, InputIconTooltip, MultiSelect, Checkbox, CheckboxList, CustomDatePicker, DatePickerRange, EmailInput, FileUpload, NumberInput, PersonalIdentityInput, Select, Textarea, TextInput, LinkArray, HiddenInput, CustomSubmitButton, CustomSubmitButtonType, RadioButton, FormLink, ClearFormButton } from "./components/Form/Form";
+﻿import { Form, Dropdown, ConditionalCheckboxList, HsaIdInput, SelectGroup, UnorderedList, TooltipItem, InputIconTooltip, MultiSelect, Checkbox, CheckboxList, CustomDatePicker, DatePickerRange, EmailInput, FileUpload, NumberInput, PersonalIdentityInput, Select, Textarea, TextInput, LinkArray, HiddenInput, CustomSubmitButton, CustomSubmitButtonType, RadioButton, FormLink, ClearFormButton } from "./components/Form/Form";
 import { Modal, ModalBody, ModalFooter, ModalSize } from "./components/Modal/Modal";
 import { Table } from "./components/Table/Table";
 import { MainNavigation } from "./components/MainNavigation/MainNavigation";
@@ -69,5 +69,6 @@ export {
     UnorderedList,
     SelectGroup,
     ConditionalCheckboxList,
-    Dropdown
+    Dropdown,
+    HsaIdInput
 }