+++ src/utils/utils.ts
@@ -1,0 +1,3 @@
+export function getNestedObjectValue(obj: any, path: string) {
+    return path.split('.').reduce((o, key) => (o && o[key] !== 'undefined' ? o[key] : undefined), obj);
+}